+++ src/lib/starter-templates/base-files/ensure-cspell.test.ts
@@ -9,6 +9,7 @@ describe('ensureCspell', () => {
     'Avenir',
     'bradlc',
     'Bstatus',
+    'bunx',
     'dbaeumer',
     'esbenp',
     'eslintcache',